+++ src/TieLn.hpp
@@ -15,7 +15,7 @@ class TieLn : public Ship {
         TieLn();
         std::string getModel() const override;
         unsigned int getModelSpeedMax() const override;
-        float getModelWeight() const override;
+        double getModelWeight() const override;
 };
